requirements.txt
requests==2.28.1
privatebinapi~=1.0.0
PyYAML~=6.0
nextcord~=2.1.0